industry2030: extract energy_general
