+++ src/climatevision/generator/business2030/b30.py
@@ -2,8 +2,6 @@
 
 from dataclasses import dataclass
 
-from ..common.g import G, GConsult
-
 from .dataclasses import (
     Vars0,
     Vars3,
@@ -23,6 +21,7 @@ from .dataclasses import (
     Vars17,
     Vars18,
 )
+from .energy_general import G, GConsult
 
 
 @dataclass(kw_only=True)
