+++ src/climatevision/generator/industry2030/__init__.py
@@ -8,7 +8,6 @@ https://localzero-generator.readthedocs.io/de/latest/sectors/industry.html
 from ..inputs import Inputs
 from ..utils import div
 from ..industry2018.i18 import I18
-from ..common.g import G, GConsult
 
 from .i30 import I30
 from .dataclasses import (
@@ -28,6 +27,7 @@ from .dataclasses import (
     Vars15,
     Vars16,
 )
+from . import energy_general
 
 
 def calc(inputs: Inputs, *, i18: I18) -> I30:
@@ -884,11 +884,9 @@ def calc(inputs: Inputs, *, i18: I18) -> I30:
         p_miner_cement.prod_volume * p_miner_cement.CO2e_combustion_based_per_t
     )
 
-    g_consult = GConsult.calc_for_industry2030(inputs)
+    general = energy_general.calc_general(inputs=inputs)
 
-    g = G.sum(g_consult)
-
-    i.demand_emplo_com = g.demand_emplo_com
+    i.demand_emplo_com = general.g.demand_emplo_com
 
     p_miner_cement.invest_per_x = ass(
         "Ass_I_P_miner_cement_kirchdorf_ratio_invest_to_prodvol_2020"
@@ -897,9 +895,9 @@ def calc(inputs: Inputs, *, i18: I18) -> I30:
 
     p_miner_cement.invest_outside = p_miner_cement.invest
 
-    i.invest_pa_com = g.invest_pa_com
+    i.invest_pa_com = general.g.invest_pa_com
     p_miner_ceram.invest_per_x = ass("Ass_I_P_miner_ceramic_ratio_invest_to_prodvol")
-    i.invest_com = g.invest_com
+    i.invest_com = general.g.invest_com
     p_miner_cement.pct_of_wage = fact("Fact_I_P_constr_civil_revenue_pct_of_wage_2018")
     p_miner_cement.invest_pa = p_miner_cement.invest / entries.m_duration_target
     p_miner_cement.cost_wage = p_miner_cement.invest_pa * p_miner_cement.pct_of_wage
@@ -1355,7 +1353,7 @@ def calc(inputs: Inputs, *, i18: I18) -> I30:
     p_metal.invest_pa_outside = (
         p_metal_steel.invest_pa_outside + p_metal_nonfe.invest_pa_outside
     )
-    i.invest = g.invest + p.invest
+    i.invest = general.g.invest + p.invest
     p_metal.invest_outside = p_metal_steel.invest_outside + p_metal_nonfe.invest_outside
     p_miner.cost_wage = (
         p_miner_cement.cost_wage
@@ -1367,8 +1365,8 @@ def calc(inputs: Inputs, *, i18: I18) -> I30:
         p_metal_steel_primary.cost_wage + p_metal_steel_secondary.cost_wage
     )
     p.demand_emplo_new = max(0, p.demand_emplo - p.emplo_existing)
-    i.demand_emplo = g.demand_emplo + p.demand_emplo
-    i.demand_emplo_new = g.demand_emplo_new + p.demand_emplo_new
+    i.demand_emplo = general.g.demand_emplo + p.demand_emplo
+    i.demand_emplo_new = general.g.demand_emplo_new + p.demand_emplo_new
     p_metal_steel.demand_electricity = (
         p_metal_steel_primary.demand_electricity
         + p_metal_steel_secondary.demand_electricity
@@ -1455,7 +1453,7 @@ def calc(inputs: Inputs, *, i18: I18) -> I30:
     p_other_further.demand_emethan = p_other_further.energy * ass(
         "Ass_I_P_other_further_fec_pct_of_gas_2050"
     )
-    i.invest_pa = g.invest_pa + p.invest_pa
+    i.invest_pa = general.g.invest_pa + p.invest_pa
     i.invest_pa_outside = p.invest_pa_outside
     p.invest_outside = (
         p_miner.invest_outside
@@ -1463,7 +1461,7 @@ def calc(inputs: Inputs, *, i18: I18) -> I30:
         + p_metal.invest_outside
         + p_other.invest_outside
     )
-    i.cost_wage = g.cost_wage + p.cost_wage
+    i.cost_wage = general.g.cost_wage + p.cost_wage
     p_other_paper.CO2e_production_based_per_t = fact(
         "Fact_I_P_other_paper_ratio_CO2e_pb_to_prodvol"
     )
@@ -1508,8 +1506,8 @@ def calc(inputs: Inputs, *, i18: I18) -> I30:
     s_renew_solarth = Vars16(energy=0)
 
     return I30(
-        g=g,
-        g_consult=g_consult,
+        g=general.g,
+        g_consult=general.g_consult,
         i=i,
         p=p,
         p_miner=p_miner,
